third_party/blink/web_tests/external/wpt/fetch/private-network-access/iframe.tentative.https.window_include=from-treat-as-public-expected.txt

This is a testharness.js-based test.
[FAIL] treat-as-public-address to local: failed preflight.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] treat-as-public-address to local: missing CORS headers.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] treat-as-public-address to local: missing PNA header.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] treat-as-public-address to private: failed preflight.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] treat-as-public-address to private: missing CORS headers.
  assert_equals: expected "timeout" but got "loaded"
[FAIL] treat-as-public-address to private: missing PNA header.
  assert_equals: expected "timeout" but got "loaded"
Harness: the test ran to completion.
